switched arena to f64 coordinates
also fixed lightcycle collision detection at speeds greater than 1 unit / physics update
also increased base lightcycle speed to 80 units / second

+++ src/server/arena/entities.rs
@@ -1,10 +1,10 @@
 use super::*;
 
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Serialize)]
+#[derive(Debug, Copy, Clone, PartialEq, Serialize)]
 pub struct Lightcycle {
     pub position: ArenaPoint,
     pub direction: Direction,
-    pub speed: isize,
+    pub speed: f64,
     pub dead: bool,
 }
 
@@ -19,7 +19,7 @@ impl Default for Lightcycle {
     }
 }
 
-#[derive(Debug, Default, Clone, Hash, PartialEq, Serialize)]
+#[derive(Debug, Default, Clone, PartialEq, Serialize)]
 pub struct Lightribbon {
     pub points: Vec<ArenaPoint>,
 }
